Add an option to force pre-O apps to use full screen aspect ratio
When an app target pre-O releases, the default max aspect ratio
is 1.86:1 which leads to ugly black areas on devices that have
screens with higher aspect ratio (for example Galaxy S8/S9).

This change adds an option to allow users to change default
aspect ratio for pre-O apps to 2.1 which would fit recent devices.

Wight554:
Simplify the code, we don't need the Settings for that, just bool

@jhenrique09 : Port to android 10

Change-Id: I32eed556a692b0a1f07cd0c9adf03eace925e136
Signed-off-by: orgesified <gesishabani1@gmail.com>